refactor: move caching logic in shared component
The component will be eventually moved to the analytics library.
